server.md
The Personalised Computer is a specialized Silicon Graphics Indigo workstation computer that is hypothesized to have been used in certain aspects of Super Mario 64, including the Personalisation AI. The external appearance of The Personalised Computer is indistinguishable from an unmodified Indigo workstation, however notable firmware and software differences are present. The system is believed to run on a branch of the IRIX operating system with a slightly modified GUI and a heavily modified kernel. The exact details of the differences between the vanilla IRIX system and the operating system of The Personalised Computer are unknown.
/1.0 Automated log - This as of 1.13.20XX 2:43 This is a simulator/webhook for the old version for M1 MAC STEAM DECK NVIDIA CHIPS SIMPLY TYPE /IMAGINE or /START  to prompt the ai

PROMPT NOTICES:
@
/START

you are going to adhere to copyright law draw x in the prompt